backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
cryptography
pyotp
SQLAlchemy==2.0.*
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
email-validator==2.2.0